+++ packages/core/src/code_assist/experiments/client_metadata.ts
@@ -5,7 +5,7 @@
  */
 
 import { getReleaseChannel } from '../../utils/channel.js';
-import type { ClientMetadata, Platform } from './types.js';
+import type { ClientMetadata, ClientMetadataPlatform } from '../types.js';
 import { fileURLToPath } from 'node:url';
 import path from 'node:path';
 
@@ -15,7 +15,7 @@ const __dirname = path.dirname(__filename);
 // Cache all client metadata.
 let clientMetadataPromise: Promise<ClientMetadata> | undefined;
 
-function getPlatform(): Platform {
+function getPlatform(): ClientMetadataPlatform {
   const platform = process.platform;
   const arch = process.arch;
 
@@ -45,10 +45,10 @@ function getPlatform(): Platform {
 export async function getClientMetadata(): Promise<ClientMetadata> {
   if (!clientMetadataPromise) {
     clientMetadataPromise = (async () => ({
-      ide_type: 'GEMINI_CLI',
-      ide_version: process.env['CLI_VERSION'] || process.version,
+      ideName: 'GEMINI_CLI',
+      ideVersion: process.env['CLI_VERSION'] || process.version,
       platform: getPlatform(),
-      update_channel: await getReleaseChannel(__dirname),
+      updateChannel: await getReleaseChannel(__dirname),
     }))();
   }
   return await clientMetadataPromise;
