+++ packages/core/src/code_assist/server.test.ts
@@ -268,7 +268,7 @@ describe('CodeAssistServer', () => {
     vi.spyOn(server, 'requestPost').mockResolvedValue(mockResponse);
 
     const metadata = {
-      ide_version: 'v0.1.0',
+      ideVersion: 'v0.1.0',
     };
     const response = await server.listExperiments(metadata);
